--- recsys.session.sql
@@ -1,1 +1,0 @@
-DELETE FROM made_recsys.authors
